# Swagger/Cargohub V2/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "CargoHub V2",
      "version": "v2"
    },
    "paths": {
...
      "/api/v2/Item_Groups": {
        "get": {
          "tags": [
            "Item_Groups"
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 1
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "post": {
          "tags": [
            "Item_Groups"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemGroup"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemGroup"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Item_Groups/{id}": {
        "get": {
          "tags": [
            "Item_Groups"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Item_Groups"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemGroup"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemGroup"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "patch": {
          "tags": [
            "Item_Groups"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              },
              "application/*+json": {
                "schema": { }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Item_Groups"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Item_Groups/{id}/items": {
        "get": {
          "tags": [
            "Item_Groups"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
      "/api/v2/Item_Lines": {
        "get": {
          "tags": [
            "Item_Lines"
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 1
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "post": {
          "tags": [
            "Item_Lines"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLine"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLine"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLine"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Item_Lines/{id}": {
        "get": {
          "tags": [
            "Item_Lines"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Item_Lines"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLine"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLine"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLine"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "patch": {
          "tags": [
            "Item_Lines"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              },
              "application/*+json": {
                "schema": { }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Item_Lines"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Item_Lines/{id}/items": {
        "get": {
          "tags": [
            "Item_Lines"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
      "/api/v2/Items/{id}": {
        "get": {
          "tags": [
            "Items"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Items"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "patch": {
          "tags": [
            "Items"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              },
              "application/*+json": {
                "schema": { }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Items"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
      "/api/v2/Items/{id}/{detailType}": {
        "get": {
          "tags": [
            "Items"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "detailType",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Item": {
          "type": "object",
          "properties": {
            "Uid": {
              "type": "string",
              "nullable": true
            },
            "Code": {
              "type": "string",
              "nullable": true
            },
            "Description": {
              "type": "string",
              "nullable": true
            },
            "Short_Description": {
              "type": "string",
              "nullable": true
            },
            "Upc_Code": {
              "type": "string",
              "nullable": true
            },
            "Model_Number": {
              "type": "string",
              "nullable": true
            },
            "Commodity_Code": {
              "type": "string",
              "nullable": true
            },
            "Item_Line": {
              "type": "integer",
              "format": "int32"
            },
            "Item_Group": {
              "type": "integer",
              "format": "int32"
            },
            "Item_Type": {
              "type": "integer",
              "format": "int32"
            },
            "Unit_Purchase_Quantity": {
              "type": "integer",
              "format": "int32"
            },
            "Unit_Order_Quantity": {
              "type": "integer",
              "format": "int32"
            },
            "Pack_Order_Quantity": {
              "type": "integer",
              "format": "int32"
            },
            "Supplier_Id": {
              "type": "integer",
              "format": "int32"
            },
            "Supplier_Code": {
              "type": "string",
              "nullable": true
            },
            "Supplier_Part_Number": {
              "type": "string",
              "nullable": true
            },
            "Created_At": {
              "type": "string",
              "nullable": true
            },
            "Updated_At": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "ItemGroup": {
          "type": "object",
          "properties": {
            "Id": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "Name": {
              "type": "string",
              "nullable": true
            },
            "Description": {
              "type": "string",
              "nullable": true
            },
            "Created_At": {
              "type": "string",
              "nullable": true
            },
            "Updated_At": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "ItemLine": {
          "type": "object",
          "properties": {
            "Id": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "Name": {
              "type": "string",
              "nullable": true
            },
            "Description": {
              "type": "string",
              "nullable": true
            },
            "Created_At": {
              "type": "string",
              "nullable": true
            },
            "Updated_At": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
      }
    }
  }